Make an apps directory, and move the variance calculator to util so we don't have a Variance directory.
